no git history (clickme for details)
I initially made this descision on march 30th '23.
At that point, the git repo took 158MB.
That's down from 165MB after gc, prune and repack.
All that for 27MB of images, 12MB being the banner.
It does not seem reasonable to me to waste this much storeage.
Neither locally, nor on githubs servers.